+++ src/main/java/net/kaczmarzyk/spring/data/jpa/web/annotation/JoinFetch.java
@@ -16,10 +16,7 @@
 package net.kaczmarzyk.spring.data.jpa.web.annotation;
 
 import javax.persistence.criteria.JoinType;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 
 /**
@@ -32,6 +29,7 @@ import java.lang.annotation.Target;
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.PARAMETER, ElementType.TYPE })
+@Repeatable(RepeatedJoinFetch.class)
 public @interface JoinFetch {
 
     String[] paths();