+++ src/test/java/net/kaczmarzyk/spring/data/jpa/web/AnnotatedSpecInterfaceWithComplexInheritanceTreeTest.java
@@ -37,13 +37,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 /**
  *  Inheritance tree of test case:
  *
- *  +--------------+                        +--------------+
- *  | @Disjunction |                        |    @Joins    |
- *  +-----^--------+                        +------^-------+
+ *  +--------------+    +--------------+    +--------------+
+ *  | @Disjunction +---->   3 x @Join  |    |    @Joins    |
+ *  +-----^--------+    +--------------+    +------^-------+
  *        |                                        |
- *  +-----+--------+                        +------+-------+
- *  |              |                        |              |
- *  +-----^--------+                        +------^-------+
+ *  +-----+--------+                        +------+-------+       +--------------+
+ *  |              |                        |              +-------> 3x@JoinFetch |
+ *  +-----^--------+                        +------^-------+       +--------------+
  *        |                                        |
  *  +-----+--------+    +--------------+    +------+-------+
  *  |   @Spec      <----+   @Spec 2    |    |              |
@@ -65,6 +65,12 @@ import static org.assertj.core.api.Assertions.assertThat;
  */
 public class AnnotatedSpecInterfaceWithComplexInheritanceTreeTest extends AnnotatedSpecInterfaceTestBase {
 
+	@Join(path = "repeatedJoin1", alias = "repeatedJoin1alias", type = LEFT, distinct = false)
+	@Join(path = "repeatedJoin2", alias = "repeatedJoin2alias", type = INNER)
+	@Join(path = "repeatedJoin3", alias = "repeatedJoin3alias", type = RIGHT, distinct = false)
+	interface FilterWithRepeatedJoinAnnotations extends Specification<Object> {
+
+	}
 
 	@Disjunction(value = {
 			@And(value = {
@@ -80,7 +86,7 @@ public class AnnotatedSpecInterfaceWithComplexInheritanceTreeTest extends Annota
 					@Spec(params = "disjunctionOr1Param1", path = "disjunctionOr1Path1", spec = NotLike.class),
 					@Spec(params = "disjunctionOr1Param2", path = "disjunctionOr1Path2", spec = NotIn.class)
 			})
-	interface DisjunctionFilter extends Specification<Object> {
+	interface DisjunctionFilter extends FilterWithRepeatedJoinAnnotations {
 	}
 
 	@Joins(
@@ -97,7 +103,14 @@ public class AnnotatedSpecInterfaceWithComplexInheritanceTreeTest extends Annota
 
 	}
 
-	interface EmptyFilterExtendingJoinsFilter extends JoinsFilter {
+	@JoinFetch(paths = {"repeatedJoinFetch1Path1", "repeatedJoinFetch1Path2"}, joinType = LEFT, distinct = false)
+	@JoinFetch(paths = {"repeatedJoinFetch2Path1"}, joinType = INNER)
+	@JoinFetch(paths = {"repeatedJoinFetch3Path1", "repeatedJoinFetch3Path2"}, joinType = RIGHT, distinct = false)
+	interface FilterWithRepeatedJoinFetchAnnotations extends Specification<Object> {
+
+	}
+
+	interface EmptyFilterExtendingJoinsFilterAndFilterWithRepeatedJoinFetch extends JoinsFilter, FilterWithRepeatedJoinFetchAnnotations {
 
 	}
 
@@ -111,7 +124,7 @@ public class AnnotatedSpecInterfaceWithComplexInheritanceTreeTest extends Annota
 	}
 
 	interface EmptyFilterExtendingEmptyFilterExtendingJoinsFilter
-			extends EmptyFilterExtendingJoinsFilter {
+			extends EmptyFilterExtendingJoinsFilterAndFilterWithRepeatedJoinFetch {
 	}
 
 	@Conjunction(
@@ -190,20 +203,20 @@ public class AnnotatedSpecInterfaceWithComplexInheritanceTreeTest extends Annota
 
 		Collection<Specification<Object>> resolvedInnerSpecs = innerSpecs(resolved);
 		assertThat(resolvedInnerSpecs)
-				.hasSize(8)
+				.hasSize(10)
 				.containsOnly(
 						// DisjunctionFilter
 						new net.kaczmarzyk.spring.data.jpa.domain.Disjunction<>(
 								new net.kaczmarzyk.spring.data.jpa.domain.Conjunction<>(
 										new Like<>(ctx.queryContext(), "disjunctionAnd1Path1", "disjunctionAnd1Param1Val"),
-										new EmptyResultOnTypeMismatch<>(new Equal<>(ctx.queryContext(), "disjunctionAnd1Path2", new String[]{"disjunctionAnd1Param2Val"}, converter))
+										new EmptyResultOnTypeMismatch<>(new Equal<>(ctx.queryContext(), "disjunctionAnd1Path2", new String[]{ "disjunctionAnd1Param2Val" }, converter))
 								),
 								new net.kaczmarzyk.spring.data.jpa.domain.Conjunction<>(
-										new EmptyResultOnTypeMismatch<>(new In<>(ctx.queryContext(), "disjunctionAnd2Path1", new String[]{"disjunctionAnd2Param1Val"}, converter)),
-										new EmptyResultOnTypeMismatch<>(new Between<>(ctx.queryContext(), "disjunctionAnd2Path2", new String[]{"disjunctionAnd2Param2Val1", "disjunctionAnd2Param2Val2"}, converter))
+										new EmptyResultOnTypeMismatch<>(new In<>(ctx.queryContext(), "disjunctionAnd2Path1", new String[]{ "disjunctionAnd2Param1Val" }, converter)),
+										new EmptyResultOnTypeMismatch<>(new Between<>(ctx.queryContext(), "disjunctionAnd2Path2", new String[]{ "disjunctionAnd2Param2Val1", "disjunctionAnd2Param2Val2" }, converter))
 								),
 								new NotLike<>(ctx.queryContext(), "disjunctionOr1Path1", "disjunctionOr1Param1Val"),
-								new EmptyResultOnTypeMismatch<>(new NotIn<>(ctx.queryContext(), "disjunctionOr1Path2", new String[]{"disjunctionOr1Param2Val"}, converter))
+								new EmptyResultOnTypeMismatch<>(new NotIn<>(ctx.queryContext(), "disjunctionOr1Path2", new String[]{ "disjunctionOr1Param2Val" }, converter))
 						),
 						// JoinsFilter
 						new net.kaczmarzyk.spring.data.jpa.domain.Conjunction<>(
@@ -225,8 +238,8 @@ public class AnnotatedSpecInterfaceWithComplexInheritanceTreeTest extends Annota
 										new EmptyResultOnTypeMismatch<>(new EqualIgnoreCase<>(ctx.queryContext(), "conjunction1or2spec1", new String[]{ "conjunction1or2spec1Val" }, converter)),
 										new EmptyResultOnTypeMismatch<>(new NotEqualIgnoreCase<>(ctx.queryContext(), "conjunction1or2spec2", new String[]{ "conjunction1or2spec2Val" }, converter))
 								),
-										new Like<>(ctx.queryContext(), "conjunction1AndSpec1", "conjunction1AndSpec1Val"),
-										new EmptyResultOnTypeMismatch<>(new Equal<>(ctx.queryContext(), "conjunction1AndSpec2", new String[] {"conjunction1AndSpec2Val"}, converter))
+								new Like<>(ctx.queryContext(), "conjunction1AndSpec1", "conjunction1AndSpec1Val"),
+								new EmptyResultOnTypeMismatch<>(new Equal<>(ctx.queryContext(), "conjunction1AndSpec2", new String[]{ "conjunction1AndSpec2Val" }, converter))
 						),
 						// OrFilter
 						new net.kaczmarzyk.spring.data.jpa.domain.Disjunction<>(
@@ -236,7 +249,19 @@ public class AnnotatedSpecInterfaceWithComplexInheritanceTreeTest extends Annota
 						// JoinFetchFilter
 						new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(new String[]{ "joinFetch1path1", "joinFetch1path2", "joinFetch1path3" }, LEFT, true),
 						// JoinFilter
-						new net.kaczmarzyk.spring.data.jpa.domain.Join<>(ctx.queryContext(), "join1", "join1alias", LEFT, false)
+						new net.kaczmarzyk.spring.data.jpa.domain.Join<>(ctx.queryContext(), "join1", "join1alias", LEFT, false),
+						// 3xJoinFetch
+						new net.kaczmarzyk.spring.data.jpa.domain.Conjunction<>(
+								new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(new String[]{"repeatedJoinFetch1Path1", "repeatedJoinFetch1Path2"}, LEFT, false),
+								new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(new String[]{"repeatedJoinFetch2Path1"}, INNER, true),
+								new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(new String[]{"repeatedJoinFetch3Path1", "repeatedJoinFetch3Path2"}, RIGHT, false)
+						),
+						// 3xJoin
+						new net.kaczmarzyk.spring.data.jpa.domain.Conjunction<>(
+								new net.kaczmarzyk.spring.data.jpa.domain.Join<>(ctx.queryContext(), "repeatedJoin1", "repeatedJoin1alias", LEFT, false),
+								new net.kaczmarzyk.spring.data.jpa.domain.Join<>(ctx.queryContext(), "repeatedJoin2", "repeatedJoin2alias", INNER, true),
+								new net.kaczmarzyk.spring.data.jpa.domain.Join<>(ctx.queryContext(), "repeatedJoin3", "repeatedJoin3alias", RIGHT, false)
+						)
 				);
 	}