+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+v2.5.0 - UNRELEASED
+======
+* Added support for repeatable `@Join` and `@JoinFetch` annotations. `@Joins` annotation is now deprecated and it's going to be removed in the future.
+
+    To specifying multiple different joins, repeated `@Join` annotation should be used: 
+    ```java
+    @RequestMapping(value = "/findBy", params = {""})
+    public void findByBadgeTypeAndOrderItemName(
+    		@Join(path = "orders", alias = "o", type = JoinType.LEFT)
+    		@Join(path = "badges", alias = "b", type = JoinType.LEFT)
+    		@Or({
+    				@Spec(path = "o.itemName", params = "order", spec = Like.class),
+    				@Spec(path = "b.badgeType", params = "badge", spec = Equal.class)
+    		}) Specification<Customer> spec) {
+    	return customerRepository.findAll(spec);
+    }
+    ```
+    instead annotation container `@Joins`
+    ```java
+    @RequestMapping(value = "/findBy", params = {""})
+    public void findByBadgeTypeAndOrderItemName(
+    		@Joins({
+    				@Join(path = "orders", alias = "o", type = JoinType.LEFT)
+    				@Join(path = "badges", alias = "b", type = JoinType.LEFT)
+    		})
+    		@Or({
+    				@Spec(path = "o.itemName", params = "order", spec = Like.class),
+    				@Spec(path = "b.badgeType", params = "badge", spec = Equal.class)
+    		}) Specification<Customer> spec) {
+    	return customerRepository.findAll(spec);
+    }
+    ```
+ 
 v2.4.1
 ======
 * Added `distinct` (default: `true`) attribute to `JoinFetch` annotation. Attribute determines that query should be distinct or not.